+++ frontend/components/hosts/HostDetails/helpers.js
@@ -4,7 +4,7 @@ const BYTES_PER_GIGABYTE = 1074000000;
 const NANOSECONDS_PER_MILLISECOND = 1000000;
 
 const inGigaBytes = (bytes) => {
-  return (bytes / BYTES_PER_GIGABYTE).toFixed(2);
+  return (bytes / BYTES_PER_GIGABYTE).toFixed(1);
 };
 
 const inMilliseconds = (nanoseconds) => {