+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+*   Update Host details cards UI
+
 *   Add support for rotating the osquery status and result log files by sending
     a SIGHUP signal to the kolide process.